+++ distribution/launcher/NOTICE
