+++ engine/launcher/src/main/scala/org/enso/launcher/cli/Arguments.scala
@@ -9,8 +9,8 @@ object Arguments {
     * [[Argument]] instance that tries to parse the String as a [[SemVer]]
     * version string.
     */
-  implicit val semverArgument: Argument[SemVer] = new Argument[SemVer] {
-    override def read(string: String): Either[List[String], SemVer] =
-      SemVer(string).toRight(List(s"`$string` is not a valid version string."))
-  }
+  implicit val semverArgument: Argument[SemVer] = (string: String) =>
+    SemVer(string).toRight(
+      List(s"`$string` is not a valid semantic version string.")
+    )
 }
