+++ engine/launcher/src/main/scala/org/enso/launcher/http/HTTPRequestBuilder.scala
@@ -1,4 +1,4 @@
-package org.enso.launcher.internal.http
+package org.enso.launcher.http
 
 import java.net.URI
