+++ engine/launcher/src/main/scala/org/enso/launcher/releases/Release.scala
@@ -49,6 +49,10 @@ trait ReleaseProvider {
 
 case class ReleaseProviderException(message: String, cause: Throwable = null)
     extends RuntimeException(message, cause) {
+
+  /**
+    * @inheritdoc
+    */
   override def toString: String =
     s"A problem occurred when trying to find the release: $message"
 }
