+++ engine/launcher/src/main/scala/org/enso/launcher/releases/github/GithubReleaseProvider.scala
@@ -1,9 +1,6 @@
 package org.enso.launcher.releases.github
 
-import java.nio.file.Path
-
-import org.enso.cli.TaskProgress
-import org.enso.launcher.releases.{Asset, Release, ReleaseProvider}
+import org.enso.launcher.releases.{Release, ReleaseProvider}
 
 import scala.util.Try
 
@@ -20,24 +17,15 @@ class GithubReleaseProvider(
 ) extends ReleaseProvider {
   private val repo = GithubAPI.Repository(owner, repositoryName)
 
+  /**
+    * @inheritdoc
+    */
   override def releaseForTag(tag: String): Try[Release] =
-    GithubAPI.getRelease(repo, tag).waitForResult().map(wrapRelease)
+    GithubAPI.getRelease(repo, tag).waitForResult().map(GithubRelease)
 
+  /**
+    * @inheritdoc
+    */
   override def listReleases(): Try[Seq[Release]] =
-    GithubAPI.listReleases(repo).map(_.map(wrapRelease))
-
-  private def wrapRelease(release: GithubAPI.Release): Release =
-    new Release {
-      override def tag: String        = release.tag
-      override def assets: Seq[Asset] = release.assets.map(wrapAsset)
-    }
-
-  private def wrapAsset(asset: GithubAPI.Asset): Asset =
-    new Asset {
-      override def fileName: String = asset.name
-      override def downloadTo(path: Path): TaskProgress[Unit] =
-        GithubAPI.downloadAsset(asset, path)
-      override def fetchAsText(): TaskProgress[String] =
-        GithubAPI.fetchTextAsset(asset)
-    }
+    GithubAPI.listReleases(repo).map(_.map(GithubRelease))
 }
